rename core folders
