+++ packages/blockchain-wallet-v4/src/network/api/index.js
@@ -1,6 +1,6 @@
-import bitcoin from './bitcoin'
+import bitcoin from './btc'
 import delegate from './delegate'
-import ethereum from './ethereum'
+import ethereum from './eth'
 import bch from './bch'
 import kvStore from './kvStore'
 import misc from './misc'
