+++ packages/blockchain-wallet-v4/src/redux/data/actionTypes.js
@@ -1,6 +1,6 @@
-import * as bitcoin from './bitcoin/actionTypes'
+import * as bitcoin from './btc/actionTypes'
 import * as coinify from './coinify/actionTypes'
-import * as ethereum from './ethereum/actionTypes'
+import * as ethereum from './eth/actionTypes'
 import * as bch from './bch/actionTypes'
 import * as misc from './misc/actionTypes'
 import * as sfox from './sfox/actionTypes'
