+++ packages/blockchain-wallet-v4/src/redux/data/actions.js
@@ -1,6 +1,6 @@
-import * as bitcoin from './bitcoin/actions'
+import * as bitcoin from './btc/actions'
 import * as coinify from './coinify/actions'
-import * as ethereum from './ethereum/actions'
+import * as ethereum from './eth/actions'
 import * as bch from './bch/actions'
 import * as misc from './misc/actions'
 import * as sfox from './sfox/actions'
