+++ packages/blockchain-wallet-v4/src/redux/data/reducers.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux'
-import bitcoin from './bitcoin/reducers'
-import ethereum from './ethereum/reducers'
+import bitcoin from './btc/reducers'
+import ethereum from './eth/reducers'
 import bch from './bch/reducers'
 import misc from './misc/reducers'
 import sfox from './sfox/reducers'
