+++ packages/blockchain-wallet-v4/src/redux/data/sagas.js
@@ -1,5 +1,5 @@
-import bitcoin from './bitcoin/sagas.js'
-import ethereum from './ethereum/sagas.js'
+import bitcoin from './btc/sagas.js'
+import ethereum from './eth/sagas.js'
 import bch from './bch/sagas.js'
 import shapeShift from './shapeShift/sagas.js'
 import sfox from './sfox/sagas.js'
