+++ packages/blockchain-wallet-v4/src/redux/data/selectors.js
@@ -1,6 +1,6 @@
-import * as bitcoin from './bitcoin/selectors'
+import * as bitcoin from './btc/selectors'
 import * as coinify from './coinify/selectors'
-import * as ethereum from './ethereum/selectors'
+import * as ethereum from './eth/selectors'
 import * as bch from './bch/selectors'
 import * as misc from './misc/selectors'
 import * as sfox from './sfox/selectors'
