+++ packages/blockchain-wallet-v4/src/redux/kvStore/actions.js
@@ -1,5 +1,5 @@
 import * as whatsNew from './whatsNew/actions'
-import * as ethereum from './ethereum/actions'
+import * as ethereum from './eth/actions'
 import * as shapeShift from './shapeShift/actions'
 import * as buySell from './buySell/actions'
 import * as contacts from './contacts/actions'
