+++ packages/blockchain-wallet-v4/src/redux/kvStore/eth/selectors.spec.js
