+++ packages/blockchain-wallet-v4/src/redux/kvStore/selectors.js
@@ -1,5 +1,5 @@
 import * as whatsNew from './whatsNew/selectors.js'
-import * as ethereum from './ethereum/selectors.js'
+import * as ethereum from './eth/selectors.js'
 import * as shapeShift from './shapeShift/selectors.js'
 import * as buySell from './buySell/selectors.js'
 import * as contacts from './contacts/selectors.js'
